+++ src/main/java/com/realtimeleaderboard/repository/GameRepository.java
@@ -5,7 +5,9 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface GameRepository extends CrudRepository<Game, Integer> {
+public interface GameRepository extends CrudRepository<Game, Long> {
     @Override
     List<Game> findAll();
+
+
 }
